+++ src/lib/storage/column_encoding_utils.cpp
@@ -1,16 +1,33 @@
-#include "utils.hpp"
+#include "column_encoding_utils.hpp"
 
 #include <boost/hana/fold.hpp>
+#include <boost/hana/pair.hpp>
+#include <boost/hana/tuple.hpp>
 #include <boost/hana/value.hpp>
 
 #include <memory>
 
-#include "encoders.hpp"
-
+#include "storage/deprecated_dictionary_column/deprecated_dictionary_encoder.hpp"
+#include "storage/dictionary_column/dictionary_encoder.hpp"
+#include "storage/run_length_column/run_length_encoder.hpp"
 #include "storage/base_value_column.hpp"
+#include "utils/enum_constant.hpp"
 #include "utils/assert.hpp"
 
 namespace opossum {
+
+namespace {
+
+/**
+ * @brief Mapping of encoding types to encoders
+ */
+constexpr auto encoder_for_type =
+    hana::make_tuple(hana::make_pair(enum_c<EncodingType, EncodingType::DeprecatedDictionary>,
+                                     hana::type_c<DeprecatedDictionaryEncoder>),
+                     hana::make_pair(enum_c<EncodingType, EncodingType::Dictionary>, hana::type_c<DictionaryEncoder>),
+                     hana::make_pair(enum_c<EncodingType, EncodingType::RunLength>, hana::type_c<RunLengthEncoder>));
+
+}  // namespace
 
 std::unique_ptr<BaseColumnEncoder> create_encoder(EncodingType encoding_type) {
   Assert(encoding_type != EncodingType::Invalid, "Encoding type must be valid.");
